# faster Component.extract()