fixes slowness in `Component.extract()` reported in https://github.com/gdsfactory/gdsfactory/issues/3086

Before
Layer 0 Time: 0.121232 s
Layer 4 Time: 31.529844 s
Layer 0 and 4 Time: 34.504553 s


After
Layer 0 Time: 0.060577 s
Layer 4 Time: 0.001984 s
Layer 0 and 4 Time: 0.002031 s

thanks @octavephotonics for reporting this!