@sebastian-goeldi 2024-08-03
For fastest performance you should use this in the recursive case https://klayout.de/doc-qt5/code/class_Cell.html#method41 and copy_shapes in the non-recursive case

@joamatab